[ALPS04357212] TEEI: Upgrade TEEI 300
1. Upgrade TEEI 300 for TEE 64bit OS and keymaster4.0.
2. Correct method to use bootprof API.

MTK-Commit-Id: 21716d4d8b8f4d4167b20400950dc53f08409faf

CR-Id: ALPS04357212
Feature: Microtrust TEEI

+++ drivers/misc/mediatek/teei/300/tz_driver/sysfs.c
@@ -573,7 +573,7 @@ static ssize_t tzdriver_dynamical_debug_store(struct device *dev,
 	if ((value == 0) || (value == 1))
 		tzdriver_dynamical_debug_flag = value;
 
-	return (ssize_t)len;
+	return len;
 }
 static DEVICE_ATTR_RW(tzdriver_dynamical_debug);
 #endif
